Fix bugs and make it finnaly work

+++ src/Definitions.h
@@ -1,9 +1,10 @@
 #pragma once
+#include <cstdint>
 
 #define NUM_THREADS 4
 typedef unsigned char BYTE;			// 0 to 255
 typedef signed short int BYTE2;		// -32768 to 32767
-typedef long LONG;
-typedef unsigned long DWORD;
-typedef unsigned short WORD;
+typedef int32_t LONG;
+typedef uint32_t DWORD;
+typedef uint16_t WORD;
 typedef struct { int x, y; } xy;
